requirements.txt
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.1
requests==2.32.3
playwright==1.44.0
web3
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
passlib==1.7.4
itsdangerous==2.2.0
